requirements.txt
discord.py
scrapy
aiohttp
ujson